requirements.txt
python-dotenv==0.21.1
aiogram==3.0.0b6
fastapi==0.89.1
